# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FlowPay Public API
  version: 1.0.0
  description: |
    API REST pública versionada bajo `/api/v1` (roadmap SC-5 / I006).
    GraphQL en `/api/graphql` es interno de la UI autenticada; no es el
    contrato de producto para integradores.
    Alias legacy sin versión (`/api/health`, `/api/cobranza/...`) delegan
    a la misma lógica; nuevos clientes deben usar `/api/v1`.
  contact:
    name: FlowPay
servers:
  - url: /
    description: Same-origin
paths:
...
  /api/v1/cobranza/importar:
    post:
      summary: Importación síncrona de cobranza
      tags: [Importaciones]
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ImportMultipart'
      responses:
        '200':
          description: Resultado de importación
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '429':
          description: Rate limit
  /api/v1/cobranza/importar/async:
    post:
      summary: Encolar importación de cobranza
      tags: [Importaciones]
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ImportMultipart'
      responses:
        '200':
          description: Job encolado
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  async:
                    type: boolean
                  job:
                    type: object
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '429':
          description: Rate limit
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        minLength: 8
        maxLength: 64
        pattern: '^[a-zA-Z0-9_-]+$'
      description: Evita jobs duplicados en reintentos (único por usuario).
  schemas:
    ImportMultipart:
      type: object
      required: [archivo, idmandante]
      properties:
        archivo:
          type: string
          format: binary
        tipo:
          type: string
          enum: [CARTERA, GESTIONES, PAGOS, PROMESAS, CONTACTOS, COMPLETO]
        idmandante:
          type: integer
        idcampana:
          type: integer
        fechaCorte:
          type: string
          format: date-time
        nombreHoja:
          type: string
        idplantillaImp:
          type: integer
  responses:
    Error:
      description: Error estable (success=false, error, code)
      content:
        application/json:
          schema:
            type: object
            required: [success, error, code]
            properties:
              success:
                type: boolean
                enum: [false]
              error:
                type: string
              code:
                type: string
              detalles:
                type: object
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth-token
...